+++ index.js
@@ -1,25 +1,50 @@
 #!/usr/bin/env node
-const commander = require('commander');
+const program = require('commander');
 const { readFileSync } = require('fs');
-
 const project = require('./cmd/project.js');
-const log = require('./util/log.js');
+const generate = require('./cmd/generate.js');
 
 const pkg = readFileSync(__dirname + '/package.json');
 
-commander
-  .version(pkg.version)
-  .usage('<keywords>')
-  .option('new [string]', 'scaffold new project in directory by name, i.e. rctr new my-app')
-  .parse(process.argv);
+program.version(pkg.version);
+
+program.command('new <name>', {executableFile: './cmd/project.js'})
+        .alias('n')
+        .description('scaffold new project')
+        .action((name) => {
+          return project({
+            name: name,
+            args: {}
+          });
+        });
+
+program.command('generate <type> <name>', {executableFile: './cmd/generate.js'})
+        .alias('g')
+        .description('generate code snippet')
+        .action((type, name) => {
+          return generate({
+            type: type,
+            name: name,
+            args: {
+              routing: program.routing,
+              lazy: program.lazy
+            }
+          });
+        });
+
+program.option('-r, --routing', 'add route for component (generate)')
+      .option('-l, --lazy', 'lazyload component (generate)')
+      .parse(process.argv);
 
 const exitHandler = (options, err) => {
-    if (err && err !== 'SIGINT') {
-        process.stdout.write('\n');
-        log.error('RCTR ERROR', err);
-        process.stdout.write('\n');
-        process.exit(1);
+    if (err === 0 || err === 1) {
+      return;
     }
+    if (err !== 'SIGINT') {
+      process.stdout.write('\n');
+      console.log(err);
+    }
+    process.exit(1);
 };
 
 // do something when app is closing
@@ -34,12 +59,4 @@ process.on('SIGUSR2', exitHandler.bind(null, { exit: true }));
 
 //catches uncaught exceptions
 process.on('uncaughtException', exitHandler.bind(null, { exit: true }));
-process.on('unhandledRejection', err => {
-  process.stdout.write('\n');
-  log.error('RCTR ERROR', err);
-  process.stdout.write('\n');
-});
-
-if (commander.new) {
-    project(commander.new);
-}
+process.on('unhandledRejection', exitHandler.bind(null, { exit: true }));