+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## 1.0.4
 
+- NEW generate command scaffolds code snippets
 - NEW changelog
+- CHANGE cli command to rx from rctr
+- FIXED issues with logs
+- FIXED issue with error reporting in project scaffold
+- FIXED issue that could cause project overwrite
 
 ## 1.0.3
 
 - FIXED async callbacks, cleaned up messy code
-- FIXED issues with logging
+- FIXED issues with logs
 - REVERT to require
 
 ## 1.0.2